refactor: 사다리게임 step2
* 사용자 이름 입력시 중복되는 값을 자동으로 제거하고 객체를 생성하도록 수정.
* player의 이름을 클래스로 분류하고 이름에 대한 검증을 해당 객체에서 수행하도록 변경.
* 기존 테스트코드에서 역할이 제거된 부분을 수정하고 신규 이름 클래스의 테스트코드 작성.
