+++ src/main/java/ladderGame/model/Player.java
@@ -4,21 +4,14 @@ import java.util.Objects;
 
 public class Player {
 
-  private static final int LIMIT_NAME_LENGTH = 5;
-
-  private static final String MSG_ERROR_FAIL_NAME_LENGTH = "사용자명은 최대 5글자까지 가능합니다.";
-
-  private final String name;
+  private final Name name;
 
   public Player(final String name) {
-    this.name = name;
-    validation(name);
+    this.name = new Name(name);
   }
 
-  private void validation(final String name) {
-    if(name.length() > LIMIT_NAME_LENGTH){
-      throw new IllegalArgumentException(MSG_ERROR_FAIL_NAME_LENGTH);
-    }
+  public String getName() {
+    return name.getName();
   }
 
   @Override
